refactor: 잘못된 image url 다루는 부분 useImageError 훅으로 분리
